fix(review): keep globe layer state coherent
